test: move into folders to reflect changes done in src/
