+++ src/tests/unit/common/providers/resolve_address/test_next_path_part.py
