+++ src/tests/unit/common/tree/mock_data/mock_blueprints/AreaWithOptionalGarden.blueprint.json
