+++ src/tests/unit/common/tree/mock_data/mock_blueprints/Box.blueprint.json
