+++ src/tests/unit/common/tree/mock_data/mock_blueprints/ChestWithOptionalBoxInside.blueprint.json
