+++ src/tests/unit/common/tree/mock_data/mock_blueprints/SpecialChildNoInherit.blueprint.json
