+++ src/tests/unit/storage/zip/test_replace_absolute_references_with_alias.py
