+++ src/tests/unit/use_cases/instantiate_entity_use_case/test_create_default_array.py
